fix bugs of comment interfaces and add getOneMonthEvents interface

+++ backend/system/src/main/java/com/powerrangers/system/modules/OrderManagement/service/impl/CommentServiceImpl.java
@@ -124,6 +124,11 @@ public class CommentServiceImpl implements CommentService {
             return new ResponseEntity<>(responseBody, HttpStatus.BAD_REQUEST);
         }
 
+        if (commentMapper.getCustomerComment(commentDTO).size() > 0) {
+            responseBody.put("error", "customer has already left a comment!");
+            return new ResponseEntity<>(responseBody, HttpStatus.BAD_REQUEST);
+        }
+
         commentDTO.setCustomerId(currUser.getId());
         commentMapper.deleteComment(commentDTO);
